requirements.txt
pymupdf
pdfplumber
python-docx
openpyxl
pytesseract
pillow
openai
sentence-transformers
faiss-cpu
pandas
tqdm
openai>=1.0.0